Removed doubled fixme. Moved random_string to test_utils.
