+++ chainstate/tx-verifier/src/transaction_verifier/utils.rs
@@ -72,7 +72,7 @@ pub fn get_output_token_id_and_amount(
             TokenData::NftIssuanceV1(_) => match include_issuance {
                 Some(tx) => {
                     let token_id = token_id(tx).ok_or(TokensError::TokenIdCantBeCalculated)?;
-                    // FIXME(nft_issuance): Find more appropriate way to check nfts
+                    // FIXME(nft_issuance): Is it okay for now? I guess, in the future PR we'll send
                     Some((CoinOrTokenId::TokenId(token_id), Amount::from_atoms(1)))
                 }
                 None => None,
